Add README.md and conf/unicorn.rb according to server config
Add README specifications and unicorn.rb configuration to enable
to run the application in production manually.

Unicorn configuration is going to be changed once capistrano task
for deployment are ready. Then the unicorn.rb configuration file
will change accordingly to capistrano deployment configuration.

In the meantime the unicorn.rb configuration has the minimum needed
to enable the application to be run in production with two
process. (just 2... small instance :P)